jobsearch_agents/jobsearch_agents/pyproject.toml
[project]
name = "GetHired"
version = "0.1.0"
description = "Job search automation ADK Agent. Designed to optimize and automate the job search process. It uses keywords from user and performs tailored job searches through google, and gathers all relevant listings, conduct research on the company, tailor resume, and prepare for application submission."
authors = [
    {name = "Aaliyah Johnson, La'Kaleigh Harris" , email = "neoaxiom@gmail.com"},
]
license = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=0.0.2)",
    # "google-adk[eval] (>=0.0.2,<1.0.0)",
    "google-genai (>=1.5.0,<2.0.0)",
    "selenium (>=4.30.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "google-cloud-bigquery (>=3.31.0,<4.0.0)",
    "absl-py (>=2.2.2,<3.0.0)",
    "google-cloud-aiplatform[agent-engines] @ git+https://github.com/googleapis/python-aiplatform.git@copybara_738852226",
    "pytest (>=8.3.5,<9.0.0)",
    "pillow (>=11.1.0,<12.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"